+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## xel.js 1.1.0  not yet released
 
+* Implement PRODUCT()
 
 ## xel.js 1.0.0  released 2024-03-17